refactor(toan-dev): format and header organization
