+++ src/main/scala/org.maogogo.moho/ServicesModule.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Maogogo Workshop
+ * Copyright 2018 maogogo organization
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,19 +17,34 @@
 package org.maogogo.moho
 
 import akka.actor.ActorSystem
+import akka.cluster.Cluster
 import akka.stream.ActorMaterializer
+import akka.util.Timeout
 import com.google.inject._
 import com.typesafe.config.{ Config, ConfigFactory }
-import com.typesafe.scalalogging.LazyLogging
 import net.codingwell.scalaguice.ScalaModule
-import org.maogogo.moho.http.RestHttpServer
+import org.maogogo.moho.cluster.SimpleClusterListener
 
-trait ServicesModule extends AbstractModule with ScalaModule {
+import scala.concurrent.duration._
 
-  override def configure(): Unit = {
-    bind[RestHttpServer]
-  }
+object ServicesModule {
+
+  lazy val systemName = "MyClusterSystem"
+
+  def apply(config: Option[Config] = None): AbstractModule with ScalaModule =
+    new AbstractModule with ScalaModule {
+      override def configure(): Unit = {
+
+        implicit val system = ActorSystem(systemName, config.getOrElse(ConfigFactory.load()))
+        bind[ActorSystem].toInstance(system)
+        bind[Config].toInstance(system.settings.config)
+        bind[ActorMaterializer].toInstance(ActorMaterializer())
+
+        bind[Cluster].toInstance(Cluster(system))
+        bind[SimpleClusterListener].in[Singleton]
+
+        bind[Timeout].toInstance(Timeout(3 seconds))
+      }
+    }
 
 }
-
-object ServicesModule extends ServicesModule
