+++ src/main/scala/org.maogogo.moho/cluster/SimpleClusterListener.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Maogogo Workshop
+ * Copyright 2018 maogogo organization
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
